+++ ESOAsg/core/eso_headers.py
@@ -1,0 +1,17 @@
+from ESOAsg import msgs
+import astropy
+
+PRODCATG = ['SCIENCE.IMAGE',
+            'SCIENCE.MEFIMAGE',
+            'SCIENCE.IMAGE.FLUXMAP',
+            'SCIENCE.SPECTRUM',
+            'SCIENCE.SRCTBL',
+            'SCIENCE.CUBE.IFS',
+            'SCIENCE.VISIBILITY',
+            'SCIENCE.CATALOG',
+            'SCIENCE.MCATALOG',
+            'SCIENCE.CATALOGTILE'
+            ]
+
+
+