+++ ESOAsg/phase3_prodcatg.py
@@ -1,0 +1,4 @@
+from ESOAsg import msgs
+from ESOAsg.datacontainers import eso_prodcatg
+
+