+++ CHANGES.rst
@@ -3,6 +3,9 @@
 
 - New HowTo on obtaining data from catalogues
 - Improved tap query to obtain ESO catalogues
+- Some adjustement for tap query to obscore
+- Starting the development of data containers
+- Working on script to make SPHERE data phase3 compliant
 
 0.1.0 (2020-10-06)
 ------------------